+++ src/math/math.ts
@@ -100,6 +100,11 @@ export class Vec2 {
         return new Vec2(Math.cos(angle), Math.sin(angle));
     }
 
+    /** Create a new Vec2 from an array */
+    static fromArray(arr: Array<number>) {
+        return new Vec2(arr[0], arr[1]);
+    }
+
     static LEFT = new Vec2(-1, 0);
     static RIGHT = new Vec2(1, 0);
     static UP = new Vec2(0, -1);
@@ -286,6 +291,10 @@ export class Vec2 {
 
     toString(): string {
         return `vec2(${this.x.toFixed(2)}, ${this.y.toFixed(2)})`;
+    }
+
+    toArray(): Array<number> {
+        return [this.x, this.y];
     }
 }